step4 refactor : 전체적인 패키지 위치 조정
