+++ src/main/kotlin/racinggame/domain/car/RandomEngineCarFactory.kt
@@ -1,4 +1,6 @@
-package racinggame.domain
+package racinggame.domain.car
+
+import racinggame.domain.engine.RandomEngine
 
 object RandomEngineCarFactory : CarFactory {
     // Todo: 임시 name
