+++ src/main/kotlin/racinggame/domain/engine/Engine.kt
@@ -1,4 +1,4 @@
-package racinggame.domain
+package racinggame.domain.engine
 
 fun interface Engine {
     fun getMovableDistance(): Int
