+++ src/main/kotlin/racinggame/domain/model/RacingStepStatus.kt
@@ -1,4 +1,4 @@
-package racinggame.model
+package racinggame.domain.model
 
 @JvmInline
 value class RacingStepStatus(val proceedLevels: List<Int>)
