+++ src/main/kotlin/racinggame/view/validator/InputValidator.kt
@@ -1,4 +1,4 @@
-package racinggame.validator
+package racinggame.view.validator
 
 interface InputValidator<T> {
     fun validate(input: T): T
