+++ src/test/kotlin/racinggame/domain/car/CarFactoryTest.kt
@@ -1,4 +1,4 @@
-package racinggame.domain
+package racinggame.domain.car
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
