+++ src/test/kotlin/racinggame/domain/engine/RandomEngineTest.kt
@@ -1,4 +1,4 @@
-package racinggame.domain
+package racinggame.domain.engine
 
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.RepeatedTest
